+++ askbot/const/__init__.py
@@ -1,10 +1,11 @@
 # encoding:utf-8
-from django.utils.translation import ugettext as _
-import re
 """
 All constants could be used in other modules
-For reasons that models, views can't have unicode text in this project, all unicode text go here.
+For reasons that models, views can't have unicode
+text in this project, all unicode text go here.
 """
+from django.utils.translation import ugettext as _
+import re
 CLOSE_REASONS = (
     (1, _('duplicate question')),
     (2, _('question is off-topic or not relevant')),
@@ -48,7 +49,8 @@ POST_SORT_METHODS = (
     ('relevance-desc', _('relevance')),
 )
 #todo: add assertion here that all sort methods are unique
-#because they are keys to the hash used in implementations of Q.run_advanced_search
+#because they are keys to the hash used in implementations
+#of Q.run_advanced_search
 
 DEFAULT_POST_SORT_METHOD = 'activity-desc'
 POST_SCOPE_LIST = (
@@ -91,28 +93,28 @@ TAG_REGEX = r'^[%s]+$' % TAG_CHARS
 TAG_SPLIT_REGEX = r'[ ,]+'
 EMAIL_REGEX = re.compile(r'\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}\b', re.I)
 
-TYPE_ACTIVITY_ASK_QUESTION=1
-TYPE_ACTIVITY_ANSWER=2
-TYPE_ACTIVITY_COMMENT_QUESTION=3
-TYPE_ACTIVITY_COMMENT_ANSWER=4
-TYPE_ACTIVITY_UPDATE_QUESTION=5
-TYPE_ACTIVITY_UPDATE_ANSWER=6
-TYPE_ACTIVITY_PRIZE=7
-TYPE_ACTIVITY_MARK_ANSWER=8
-TYPE_ACTIVITY_VOTE_UP=9
-TYPE_ACTIVITY_VOTE_DOWN=10
-TYPE_ACTIVITY_CANCEL_VOTE=11
-TYPE_ACTIVITY_DELETE_QUESTION=12
-TYPE_ACTIVITY_DELETE_ANSWER=13
-TYPE_ACTIVITY_MARK_OFFENSIVE=14
-TYPE_ACTIVITY_UPDATE_TAGS=15
-TYPE_ACTIVITY_FAVORITE=16
+TYPE_ACTIVITY_ASK_QUESTION = 1
+TYPE_ACTIVITY_ANSWER = 2
+TYPE_ACTIVITY_COMMENT_QUESTION = 3
+TYPE_ACTIVITY_COMMENT_ANSWER = 4
+TYPE_ACTIVITY_UPDATE_QUESTION = 5
+TYPE_ACTIVITY_UPDATE_ANSWER = 6
+TYPE_ACTIVITY_PRIZE = 7
+TYPE_ACTIVITY_MARK_ANSWER = 8
+TYPE_ACTIVITY_VOTE_UP = 9
+TYPE_ACTIVITY_VOTE_DOWN = 10
+TYPE_ACTIVITY_CANCEL_VOTE = 11
+TYPE_ACTIVITY_DELETE_QUESTION = 12
+TYPE_ACTIVITY_DELETE_ANSWER = 13
+TYPE_ACTIVITY_MARK_OFFENSIVE = 14
+TYPE_ACTIVITY_UPDATE_TAGS = 15
+TYPE_ACTIVITY_FAVORITE = 16
 TYPE_ACTIVITY_USER_FULL_UPDATED = 17
 TYPE_ACTIVITY_EMAIL_UPDATE_SENT = 18
 TYPE_ACTIVITY_MENTION = 19
 TYPE_ACTIVITY_UNANSWERED_REMINDER_SENT = 20
-#TYPE_ACTIVITY_EDIT_QUESTION=17
-#TYPE_ACTIVITY_EDIT_ANSWER=18
+#TYPE_ACTIVITY_EDIT_QUESTION = 17
+#TYPE_ACTIVITY_EDIT_ANSWER = 18
 
 #todo: rename this to TYPE_ACTIVITY_CHOICES
 TYPE_ACTIVITY = (
@@ -250,11 +252,22 @@ DEFAULT_USER_STATUS = 'w'
 #number of items to show in user views
 USER_VIEW_DATA_SIZE = 50
 
+#not really dependency, but external links, which it would 
+#be nice to test for correctness from time to time
 DEPENDENCY_URLS = {
-    'mathjax': 'http://www.mathjax.org/resources/docs/?installation.html',
+    'akismet': 'https://akismet.com/signup/',
+    'cc-by-sa': 'http://creativecommons.org/licenses/by-sa/3.0/legalcode',
+    'embedding-video': \
+        'http://askbot.org/doc/optional-modules.html#embedding-video',
     'favicon': 'http://en.wikipedia.org/wiki/Favicon',
-    'embedding-video': 'http://askbot.org/doc/optional-modules.html#embedding-video',
-    'cc-by-sa': 'http://creativecommons.org/licenses/by-sa/3.0/legalcode'
+    'facebook-apps': 'http://www.facebook.com/developers/createapp.php',
+    'google-webmaster-tools': 'https://www.google.com/webmasters/tools/home',
+    'identica-apps': 'http://identi.ca/settings/oauthapps',
+    'noscript': 'https://www.google.com/support/bin/answer.py?answer=23852',
+    'linkedin-apps': 'https://www.linkedin.com/secure/developer',
+    'mathjax': 'http://www.mathjax.org/resources/docs/?installation.html',
+    'recaptcha': 'http://google.com/recaptcha',
+    'twitter-apps': 'http://dev.twitter.com/apps/',
 }
 
 PASSWORD_MIN_LENGTH = 8